* main: (416 commits)
  Set current year as default
  Bump eslint-config-next from 13.2.0 to 13.2.1 (#387)
  Bump eslint-config-next from 13.1.6 to 13.2.0 (#386)
  Bump @typescript-eslint/eslint-plugin from 5.52.0 to 5.53.0 (#385)
  Bump @mui/material from 5.11.9 to 5.11.10 (#384)
  Bump @mui/lab from 5.0.0-alpha.114 to 5.0.0-alpha.120 (#383)
  Remove console.error
  Remove console and add eslint rule
  Improve history diagram
  Reorganize diagrams
  Move done filtering to analysis
  Fix import
  Simplify diagram cards
  Improve analysis card design
  Make sure ratio has same order
  Add days overview
  Add year filter
  Make diagrams responsive
  Move DiagramRow to analysis
  Add titles to diagrams
  ...

+++ lib/application/ExportImportApplicationService.ts
@@ -1,21 +1,36 @@
-import TimeLog from "../domain/TimeLog";
 import TimeLogRepository from "../infrastructure/TimeLogRepository";
+import TimeLogsFile from "../infrastructure/TimeLogsFile";
+import { format, isAfter } from "date-fns";
+import { todayWorkdayEnd } from "../domain/time-constants";
 
 export default class ExportImportApplicationService {
-    private readonly timeLogRepository: TimeLogRepository
+  private readonly timeLogRepository: TimeLogRepository;
 
-    constructor() {
-        this.timeLogRepository = new TimeLogRepository()
-    }
+  constructor() {
+    this.timeLogRepository = new TimeLogRepository();
+  }
 
-    async getAllTimeLogs() {
-        return await this.timeLogRepository.all()
+  async importFile(file: File) {
+    const timeLogs = await new TimeLogsFile().read(file);
+    if (timeLogs.length > 0) {
+      await this.timeLogRepository.deleteAll();
+      await this.timeLogRepository.saveAll(timeLogs);
     }
+  }
 
-    async import(timeLogs: TimeLog[]) {
-        await this.timeLogRepository.deleteAll()
-        await this.timeLogRepository.saveAll(timeLogs)
+  async exportAllTimeLogs() {
+    const timeLogs = await this.timeLogRepository.all();
+    const fileName = `${format(new Date(), "yyyy-MM-dd_HH-mm")}_overtimer.json`;
+    await new TimeLogsFile().write(fileName, timeLogs);
+  }
+
+  async exportAllTimeLogsAfterWorkdayEnd(date: Date) {
+    const afterWorkday = isAfter(date, todayWorkdayEnd());
+    if (afterWorkday) {
+      await this.exportAllTimeLogs();
     }
+  }
 }
 
-export const exportImportApplicationService = new ExportImportApplicationService()
+export const exportImportApplicationService =
+  new ExportImportApplicationService();
